Version 5.00 (04/2009)

+++ asxxsrc/asexpr.c
@@ -1,8 +1,21 @@
 /* asexpr.c */
 
 /*
- * (C) Copyright 1989-2006
- * All Rights Reserved
+ *  Copyright (C) 1989-2009  Alan R. Baldwin
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
  *
  * Alan R. Baldwin
  * 721 Berkeley St.
@@ -355,7 +368,7 @@ struct expr *esp;
 	}
 	if (c == '\"') {
 		esp->e_mode = S_USER;
-		if (hilo) {
+		if ((int) hilo) {
 		    esp->e_addr  = (getmap(-1)&0377)<<8;
 		    esp->e_addr |= (getmap(-1)&0377);
 		} else {
